refactor: remove internal/layout/Engine + update CHANGELOG (ADR-032 CACHE-030)
Engine (centralized layout cache) has 0 production usage — confirmed by
codebase audit. Per-widget layout caching is now handled by
widget.LayoutChild on WidgetBase (ADR-032 Phase 2b, #160).

- Delete engine.go + engine_test.go (-769 LOC)
- Extract Layoutable interface to layoutable.go (used by Flex/Grid/Stack)
- Extract mockLayoutable to mock_test.go
- Remove Engine cache benchmarks from bench_test.go
- Update doc.go and CHANGELOG
